+++ RPI4/include/SensorHub.hpp
@@ -27,8 +27,8 @@ public:
     const static int NUMENC = 6;    
     const static int NUMPRE = 8; //always 8 since ADC has 8 channels
     
-    const float ENC_DEN = 4096.0;
-    const float PRE_DEN = 65536;
+    const double ENC_DEN = 4096.0;
+    const double PRE_DEN = 65536;
 
     static SensorHub& GetInstance();
     static const std::array<double,NUMENC>& GetEncData(); //I did not use lock here since they will be read-only arrays
